Moved calculator functions
